Created the Homing Rose Petals, and Started Attacks
When Ruby Rose dodges while being powered up with dust, then she will
leave a homing rose petal in her wake that can cause DOT, but not base.
Also knocks back the player a bit via the direction of the rose petal
movement

Started working on the hit trace detection for multiplayer melee
attacks. Right now animation is playing on both client and server, same
with Line Traces. Server is still the one that uses the hit data. Only
server will say if you hit someone or not. NOT the client