2D Shooting
All characters will be able to shoot in 2D
the mouse is where their character will aim to.

Right now the Aiming is working, new animations, and other mechanics are
colliding with this in a problematic way..

The Server is the only one working right now, needs a client method to
find the Mouse's position then on the server set the angle in the server
method. will do tomorrow