Wall Jumping system bugs
The biggest one right now is the client will always try to hang on the
air, while the server sees the client falling normally.

This causes the two to conflict, and lock out player control and causes
some wierd issues. need to get them back in sync.

Sometimes the flying state isn't removed and then the character flys
aways when jumping, However after a editor reset it works again. This
might have to do with what is broken above.

The client seems to have trouble replicating variables. might need to
create methods to force the client, and server to be in sync