In case i forget
The way to fix the Wall slide bug is to move the character out of the
Wall trace zone (diagonally) when they let go of a ledge. This will
force the player to go toward the wall if they want to wall jump, but
can free fall down if they want that too.