More Bugs
Fixed Multiplayer Syncing with server issues with the wall jumping
system.

The advance techs aren't breaking the game anymore.
Neither will a busy CPU. Seems to be "fixed" for now

However the only issue that is left is the wall slide, and the wall
climb through a celiing.

Fixing this tomorrow by setting an offset to the let go feature, also
cast a ray above the character to test for collision which will not
allow a player to hang from the wall no mater what.