Multiplayer animations!
The ruby rose class can now fight in a full on  multiplayer setting. The
animations are replicated with Multicast on the click, and will either
run on ther server, or the multicast to all the clients if the server is
playing. All screens can see the animations, and even take damage from
them.
Will be adding issues for the bugs i know of soon as well!